requirements.txt
pyyaml
django